% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_entropy.R
\name{Get_entropy}
\alias{Get_entropy}
\title{Get_entropy function}
\usage{
Get_entropy(data, K, min_exp = 0.05)
}
\arguments{
\item{data}{Enter the gene expression data, genes*cells}

\item{K}{Enter the number of PCs eg. 2, 1~5 are recommended.}

\item{min_exp}{Enter the minimum expression of genes,eg.0.05}
}
\description{
This function allows you to calculate the intrinsic entropy, extrinsic entropy and total entropy of genes based on date.
We assumed that the data follows Gauss distribition.
}
\examples{
Get_entropy(data,1,0.05)

}
\keyword{Get_entropy}
